Rename include directory to lions from lions_os
