+++ components/micropython/fs_helpers.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include <assert.h>
 #include <stdio.h>
-#include <lions_os/fs/protocol.h>
+#include <lions/fs/protocol.h>
 #include "micropython.h"
 #include "fs_helpers.h"
